+++ CHANGELOG.md
@@ -4,6 +4,9 @@
   - `xtream_skip_live_direct_source`
   - `xtream_skip_video_direct_source`
 * internal optimization/refactoring to avoid string cloning.
+* new options for downloading media files from web-ui 
+  - `organize_into_directories`
+  - `episode_pattern`
 
 # v1.1.2(2023-11-03)
 * Fixed epg for xtream